refactor(web): merge stores/ into store/ directory [tier-c]
The auth store was the sole occupant of src/stores/, split from the other
Zustand stores in src/store/ (banking, budgets, transactions). This was a
residue of an unfinished rename — src/store/index.ts:51 already had a
commented-out export for auth.store. Consolidated into single directory
with consistent dot-separator naming convention.

12 imports updated, auth-store.ts renamed to auth.store.ts.
