+++ apps/web/__tests__/components/auth/protected-route.test.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, waitFor } from '../../utils/test-utils';
 import { ProtectedRoute, withAuth } from '../../../src/components/auth/protected-route';
-import { useAuthStore } from '../../../src/stores/auth-store';
+import { useAuthStore } from '../../../src/store/auth.store';
 
 // Mock next/navigation
 const mockPush = vi.fn();
@@ -20,7 +20,7 @@ vi.mock('next/navigation', () => ({
 }));
 
 // Mock auth store
-vi.mock('../../../src/stores/auth-store', () => ({
+vi.mock('../../../src/store/auth.store', () => ({
   useAuthStore: vi.fn(),
 }));
